@@ -16,16 +16,17 @@
   },
   "devDependencies": {
     "babel": "^5.8.23",
-    "babel-eslint": "^4.1.3",
+    "babel-eslint": "^10.0.1",
     "chai": "^3.5.0",
-    "coveralls": "^2.11.12",
-    "eslint": "^1.10.3",
+    "coveralls": "^3.0.3",
+    "eslint": "^5.16.0",
     "eslint-plugin-react": "^3.15.0",
     "gulp": "^3.9.1",
     "isomorphic-fetch": "^2.2.1",
     "istanbul": "^0.4.5",
     "jsdom": "^9.5.0",
-    "mocha": "^3.0.2",
+    "mocha": "^6.1.4",
+    "prop-types": "^15.7.1",
     "react": "^15.0",
     "react-addons-shallow-compare": "^15.0",
     "react-addons-test-utils": "^15.0",
