@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import Select from 'react-select';
 import Gravatar from 'react-gravatar';
 
@@ -7,14 +8,14 @@ const GRAVATAR_SIZE = 15;
 
 const GravatarOption = React.createClass({
 	propTypes: {
-		children: React.PropTypes.node,
-		className: React.PropTypes.string,
-		isDisabled: React.PropTypes.bool,
-		isFocused: React.PropTypes.bool,
-		isSelected: React.PropTypes.bool,
-		onFocus: React.PropTypes.func,
-		onSelect: React.PropTypes.func,
-		option: React.PropTypes.object.isRequired,
+		children: PropTypes.node,
+		className: PropTypes.string,
+		isDisabled: PropTypes.bool,
+		isFocused: PropTypes.bool,
+		isSelected: PropTypes.bool,
+		onFocus: PropTypes.func,
+		onSelect: PropTypes.func,
+		option: PropTypes.object.isRequired,
 	},
 	handleMouseDown (event) {
 		event.preventDefault();
@@ -52,9 +53,9 @@ const GravatarOption = React.createClass({
 
 const GravatarValue = React.createClass({
 	propTypes: {
-		children: React.PropTypes.node,
-		placeholder: React.PropTypes.string,
-		value: React.PropTypes.object
+		children: PropTypes.node,
+		placeholder: PropTypes.string,
+		value: PropTypes.object
 	},
 	render () {
 		var gravatarStyle = {
@@ -78,8 +79,8 @@ const GravatarValue = React.createClass({
 
 const UsersField = React.createClass({
 	propTypes: {
-		hint: React.PropTypes.string,
-		label: React.PropTypes.string,
+		hint: PropTypes.string,
+		label: PropTypes.string,
 	},
 	getInitialState () {
 		return {};
