@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
 const Value = React.createClass({
@@ -6,12 +7,12 @@ const Value = React.createClass({
 	displayName: 'Value',
 
 	propTypes: {
-		children: React.PropTypes.node,
-		disabled: React.PropTypes.bool,               // disabled prop passed to ReactSelect
-		id: React.PropTypes.string,                   // Unique id for the value - used for aria
-		onClick: React.PropTypes.func,                // method to handle click on value label
-		onRemove: React.PropTypes.func,               // method to handle removal of the value
-		value: React.PropTypes.object.isRequired,     // the option object for this value
+		children: PropTypes.node,
+		disabled: PropTypes.bool,               // disabled prop passed to ReactSelect
+		id: PropTypes.string,                   // Unique id for the value - used for aria
+		onClick: PropTypes.func,                // method to handle click on value label
+		onRemove: PropTypes.func,               // method to handle removal of the value
+		value: PropTypes.object.isRequired,     // the option object for this value
 	},
 
 	handleMouseDown (event) {
